TinderBot/requirements.txt
selenium==4.7.2
undetected-chromedriver==3.2.1
webdriver-manager==3.5.4
Pillow==9.3.0
deepface
opencv-python
numpy
openai
python-dotenv
selenium-stealth
geckodriver-autoinstaller
